New version
This change the MQTT broker easily.
Also, by reading the MHDR of the LoRaWAN message, I can identify the message type (label).
I also modified it so that on Windows it saves the data into a new CSV file, allowing us to calculate the number of join requests and uplinks, similar to what we used to test Amir’s detection method.